Added new module presets and new class PresetsSegmentation. Simplified the API of the segmentationLUT (now you can only set and get the segmentation (which contains all the information). The presets segmentation allows you to add presets, and get and set presets.
