+++ src/presets/presets.js
@@ -1,0 +1,8 @@
+import Segmentation from './presets.segmentation';
+/**
+ * @module presets
+ */
+
+export default {
+  Segmentation,
+};
